Fixing unit tests
