+++ app/actions/note.js
@@ -1,0 +1,8 @@
+// @flow
+export const NOTE_SAVE = 'NOTE_SAVE';
+
+export function save() {
+  return {
+    type: NOTE_SAVE
+  };
+}
