+++ app/reducers/index.js
@@ -1,10 +1,10 @@
 // @flow
 import { combineReducers } from 'redux';
 import { routerReducer as routing } from 'react-router-redux';
-import counter from './counter';
+import note from './note';
 
 const rootReducer = combineReducers({
-  counter,
+  note,
   routing
 });
