+++ test/actions/note.spec.js
@@ -1,0 +1,10 @@
+/* eslint-disable no-unused-expressions */
+import { expect } from 'chai';
+import * as actions from '../../app/actions/note';
+
+
+describe('actions', () => {
+  it('should save should create NOTE_SAVE action', () => {
+    expect(actions.save()).to.deep.equal({ type: actions.NOTE_SAVE });
+  });
+});
